Overhaul for upstream-reproducible tarballs
This change makes it possible to generate primary and testdata
tarballs that match Google's lite and testdata tarballs bit-perfectly.
First confirmed with the 132.0.6834.110 release.

In the GitHub workflow, check that the third_party/{llvm,rust-src}
directories are present in the tarball tree, or else put a prominent
warning on the workflow run summary page. Also, reduce Git noise by
setting advice.detachedHead=false.

New file `check.patch`: This patch is not meant to modify any files;
it is used only to verify that a critical section of the upstream
`publish_tarball.py` script hasn't changed on us. If it causes a
`patch --dry-run` invocation to fail, then that indicates an update to
`package_chromium.sh` is probably needed.

Delete `export_tarball.py`: We now pull down the upstream Git repo
that contains this script, and run it there.

Numerous changes in `package_chromium.sh`:

* Miscellaneous updates to comment text

* Rework the `get_depot_tools()` function into `get_google_repo()`, so
  that we can use it to pull down both the `depot_tools` and `build`
  repositories.

* More consistent handling of function arguments

* Reformatting of various commands for better legibility, as well as
  more generous use of whitespace.

* Add run hooks for LLVM and Rust elements. The bulk of this is not
  used in our tarball, but we do need various bits from it to be able
  to match what Google has. Run these in a conditional so that their
  failure doesn't stop the tarball-generation process, since they have
  been known to break due to Git conflicts.

* Also add run hook to copy in `clang-format`.

* New function `prune_lite_excluded_dirs()` that matches what the
  `export_lite_tarball()` function in `publish_tarball.py` does (and
  the reason why we have `check.patch`, since we need to keep this in
  sync)

* Invoke the upstream `export_tarball.py` script in situ instead of
  our own copy.

* Invoke the upstream `generate_hashes.py` script instead of our
  re-creation.

* Use `check.patch` to check for breaking changes in the upstream
  scripts.

* Temporarily patch scripts in the `build` repo and main source tree
  for our purposes (patches described below).

* Set an environment variable, implemented by one of the aforemention
  patches, to reduce the size of the LLVM and Rust clone operations.

* Prune the source tree for lite-ness prior to exporting tarballs.

New file `tweak-src.patch`: Patches `src/tools/clang/scripts/build.py`
in the main source tree to implement `EXTRA_GIT_CLONE_ARGS` and skip
the downloading of the Debian sysroots, all to reduce the size of the
LLVM/Rust Git pulls.

New file `tweak-tools.patch`: Patch the upstream `export_tarball.py`
script with a hack to make the `tarfile` module produce bit-identical
output to Google's.