CI: add manual tag workflow
I'm doing some work downstream where I need tarballs for arbitrary
versions of Chromium. This seems like it may also be useful to others,
so provide the ability to create tarballs and releases for abritrary
tags, using some webrequests to ensure that the tag actually exists
before creating the tag and running other workflows.

The reason for creation is reported in the IRC notification.

Signed-off-by: Matt Jolly <kangie@gentoo.org>